test(ops): move market ranking charter guard to static contract path
