+++ tests/webui/test_market_dashboard_readonly_structure_contract_v0.py
@@ -743,47 +743,3 @@ def test_double_play_market_dashboard_excludes_run_projection_landmark_v0(
 
     assert "market-v0-landmark-run-projection-h2" not in html
     assert 'data-market-v0-run-projection="true"' not in html
-
-
-def test_market_surface_ranking_funnel_producer_charter_v0() -> None:
-    """Market Ranking Funnel Producer v0 charter stays read-only and non-authorizing."""
-    market_surface_doc = "docs/webui/MARKET_SURFACE_V0.md"
-    surface = (project_root / market_surface_doc).read_text(encoding="utf-8")
-
-    required_tokens = [
-        "### Market Ranking Funnel Producer v0 (read-only charter)",
-        "market_ranking_funnel_readmodel.v0",
-        "PEAK_TRADE_MARKET_RANKING_FUNNEL_ENABLED=1",
-        "GET &#47;market",
-        "&#47;market&#47;double-play",
-        "&#47;api&#47;market&#47;ranking",
-        "endpoint in v0",
-        "tests/webui/test_market_dashboard_readonly_structure_contract_v0.py",
-        "DASHBOARD_AUTHORITY_CHANGED=false",
-        "RANKING_PRODUCER_AUTHORIZES_TRADES=false",
-        "FuturesRankingSnapshot",
-        "must not be directly wired",
-        "non_authorizing=true",
-        "universe",
-        "shortlist",
-        "selected",
-    ]
-
-    for token in required_tokens:
-        assert token in surface
-
-    forbidden_promotions = [
-        "trade signal",
-        "approval",
-        "readiness",
-        "live authorization",
-        "strategy activation",
-        "Master V2",
-        "Double Play trading input",
-    ]
-
-    forbidden_section_start = surface.index("Forbidden promotions:")
-    forbidden_section = surface[forbidden_section_start:]
-
-    for token in forbidden_promotions:
-        assert token in forbidden_section
